lpm6: fix type of field next_hop of struct rte_lpm6_rule
The parameter next_hop of a rule in an LPM6 table is 21 bits long
according to struct rte_lpm6_tbl_entry, but struct rte_lpm6_rule
had only 8 bits for it. This would truncate the value of next_hop
on a couple of occasions.

+++ lib/librte_lpm/rte_lpm6.h
@@ -25,7 +25,7 @@ extern "C" {
 /** Rules tbl entry structure. */
 struct rte_lpm6_rule {
 	uint8_t ip[RTE_LPM6_IPV6_ADDR_SIZE]; /**< Rule IP address. */
-	uint8_t next_hop; /**< Rule next hop. */
+	uint32_t next_hop; /**< Rule next hop. */
 	uint8_t depth; /**< Rule depth. */
 };
